Update README.md
In this code, I started by using the "array" command to list out the songs in a particular order; the name of the song, the artist, and the number of streams for that specific song, which I listed off of Spotify. After doing this, I then used the print method to display the list of songs in that order. I then added a "choices" method so that it would show up immediately on the screen and included the following options; "Add a song", "Delete a song", "Print songs", "Print Popular Songs", and "Stop", (which would end the program). The way in which a song would be added was coded by using the format in the array, "song, artist, streams", and the print method was used so that it would display the values. The method for deleting a song was coded by using the array.remove() command and the if method was used to ensure that the value was equal to what was inputted before executing. The printSongs method was used to display the songs in a drop-down list, and the layout for each song and its information are coded to be in that order. For printing popular songs, the streams (int) had to be above a certain number, so the if method was used so that the numbers were compared and if it was higher, the song would be printed. The "else if" statement was used several times so that alternative routes would be executed depending on the user input.